refactor(heartbeat): move task body renderers into bounded module
